+++ packages/auth/src/Auth.ts
@@ -88,7 +88,8 @@ export default class AuthClass {
             identityPoolId, 
             mandatorySignIn,
             refreshHandlers,
-            storage
+            storage,
+            identityPoolRegion
         } = this._config;
 
         if (!this._config.storage) {
@@ -118,7 +119,7 @@ export default class AuthClass {
 
         Credentials.configure({
             mandatorySignIn,
-            region,
+            region: identityPoolRegion || region,
             userPoolId,
             identityPoolId,
             refreshHandlers,
@@ -403,6 +404,8 @@ export default class AuthClass {
 
     /**
      * get user current preferred mfa option
+     * this method doesn't work with totp, we need to deprecate it.
+     * @deprecated
      * @param {CognitoUser} user - the current user
      * @return - A promise resolves the current preferred mfa option if success
      */
@@ -418,6 +421,23 @@ export default class AuthClass {
             });
         });
     }
+
+    /**
+     * get preferred mfa method
+     * @param {CognitoUser} user - the current cognito user
+     */
+    public getPreferredMFA(user: any): Promise<string> {
+        return new Promise((res, rej) => {
+            user.getUserData((err, data) => {
+                if (err) {
+                    logger.debug('getting preferred mfa failed', err);
+                    rej('getting preferred mfa failed: ' + err);
+                }
+                const preferredMFA = data.PreferredMfaSetting || 'NOMFA';
+                res(preferredMFA);
+            });
+        });
+    }
     
     /**
      * set preferred MFA method
@@ -425,8 +445,9 @@ export default class AuthClass {
      * @param {string} mfaMethod - preferred mfa method
      * @return - A promise resolve if success
      */
-    public setPreferredMFA(user : any, mfaMethod : string): Promise<any> {
-        let smsMfaSettings = null;
+    public async setPreferredMFA(user : any, mfaMethod : string): Promise<any> {
+        let smsMfaSettings = await this.getPreferredMFA(user) === 'SMS_MFA'?
+            { PreferredMfa : false, Enabled : false } : null;
         let totpMfaSettings = {
             PreferredMfa : false,
             Enabled : false