+++ packages/aws-amplify-angular/src/__tests__/components/authenticator/confirm-sign-up.component.core.spec.ts
@@ -2,17 +2,17 @@ import { Component } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { BrowserDynamicTestingModule, platformBrowserDynamicTesting } from '@angular/platform-browser-dynamic/testing';
 import { AmplifyService } from '../../../providers/amplify.service'
-import { ConfirmSignInComponent } from '../../../components/authenticator/confirm-sign-in.component'
+import { ConfirmSignUpComponentCore } from '../../../components/authenticator/confirm-sign-up-component/confirm-sign-up.component.core'
 
 
-describe('ConfirmSignInComponent: ', () => {
+describe('ConfirmSignUpComponentCore: ', () => {
 
-  let component: ConfirmSignInComponent;
+  let component: ConfirmSignUpComponentCore;
   let service: AmplifyService;
 
   beforeEach(() => { 
     service = new AmplifyService();
-    component = new ConfirmSignInComponent(service);
+    component = new ConfirmSignUpComponentCore(service);
   });
 
   afterEach(() => {
@@ -25,30 +25,24 @@ describe('ConfirmSignInComponent: ', () => {
     expect(component).toBeTruthy();
   });
 
-  it('...should have a setCode method', () => {
-    expect(component.setCode).toBeTruthy();
-  });
-
-
-  it('...should have a code property that is initally undefined', () => {
-    expect(component.code).toBeUndefined();
-  })
-
-  it('...the setCode method should set the component\'s code property', () => {
-    component.setCode('200');
-    expect(component.code).toEqual('200');
-  });
-
   it('...should have an onConfirm method', () => {
     expect(component.onConfirm).toBeTruthy();
   });
 
+  it('...should have an onResend method', () => {
+    expect(component.onResend).toBeTruthy();
+  }); 
+
   it('...should have an onSignIn method', () => {
     expect(component.onSignIn).toBeTruthy();
+  }); 
+
+  it('...should have an setCode method', () => {
+    expect(component.setCode).toBeTruthy();
   });
 
-  it('...should have a _setError method', () => {
-    expect(component._setError).toBeTruthy();
+  it('...should have an setUsername method', () => {
+    expect(component.setUsername).toBeTruthy();
   });
 
 });